Make wrapper and update manager support both 4.x and 5.x, native and generic

+++ other/wrapper/nzbhydra2wrapperPy3.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-
+import json
 import sys
 
 CURRENT_PYTHON = sys.version_info[:2]
@@ -20,6 +20,19 @@ import shutil
 import subprocess
 import zipfile
 from logging.handlers import RotatingFileHandler
+from enum import Enum
+
+
+class ReleaseType(str, Enum):
+    NATIVE = "native"
+    GENERIC = "generic"
+
+
+class OsType(str, Enum):
+    WINDOWS = "windows"
+    LINUX = "linux"
+    GENERIC = "generic"
+
 
 jarFile = None
 basepath = None
@@ -153,8 +166,9 @@ def update():
     global jarFile
     basePath = getBasePath()
     updateFolder = os.path.join(args.datafolder, "update")
+    releaseType = determineReleaseType()
     libFolder = os.path.join(basePath, "lib")
-    isWindows = any([x for x in os.listdir(basePath) if x.lower().endswith(".exe")])
+    isWindows = platform.system().lower() == "windows"
     logger.debug("Is Windows installation: %r", isWindows)
     if not os.path.exists(updateFolder):
         logger.critical("Error: Update folder %s does not exist", updateFolder)
@@ -169,7 +183,7 @@ def update():
         with zipfile.ZipFile(updateZip, "r") as zf:
             logger.info("Extracting updated files to %s", basePath)
             for member in zf.namelist():
-                if not member.lower() == "nzbhydra2" and not member.lower().endswith(".exe"):
+                if (not member.lower() == "nzbhydra2" and not member.lower().endswith(".exe")) or member.lower() == "core.exe":
                     logger.debug("Extracting %s to %s", member, basePath)
                     try:
                         zf.extract(member, basePath)
@@ -178,19 +192,22 @@ def update():
                         sys.exit(-2)
         logger.info("Removing update ZIP %s", updateZip)
         os.remove(updateZip)
-        filesInLibFolder = [f for f in os.listdir(libFolder) if os.path.isfile(os.path.join(libFolder, f)) and f.endswith(".jar")]
-        logger.info("Found %d JAR files in lib folder", len(filesInLibFolder))
-        for file in filesInLibFolder:
-            logger.info("Found file: %s", file)
-        if len(filesInLibFolder) == 2:
-            logger.info("Deleting old JAR %s", jarFile)
-            os.remove(jarFile)
-        elif len(filesInLibFolder) == 1:
-            if filesInLibFolder[0] == os.path.basename(jarFile):
-                logger.warning("New JAR file in lib folder is the same as the old one. The update may not have found a newer version or failed for some reason")
+        if releaseType == ReleaseType.GENERIC:
+            logger.info("Updating lib folder for generic release type")
+            filesInLibFolder = [f for f in os.listdir(libFolder) if os.path.isfile(os.path.join(libFolder, f)) and f.endswith(".jar")]
+            logger.info("Found %d JAR files in lib folder", len(filesInLibFolder))
+            for file in filesInLibFolder:
+                logger.info("Found file: %s", file)
+            if len(filesInLibFolder) == 2:
+                logger.info("Deleting old JAR %s", jarFile)
+                os.remove(jarFile)
+            elif len(filesInLibFolder) == 1:
+                if filesInLibFolder[0] == os.path.basename(jarFile):
+                    logger.warning("New JAR file in lib folder is the same as the old one. The update may not have found a newer version or failed for some reason")
+            else:
+                logger.warning("Expected the number of JAR files in folder %s to be 2 but it's %d. This will be fixed with the next start", libFolder, len(filesInLibFolder))
         else:
-            logger.warning("Expected the number of JAR files in folder %s to be 2 but it's %d. This will be fixed with the next start", libFolder, len(filesInLibFolder))
-
+            logger.info("Skipping lib folder for native release type")
     except zipfile.BadZipfile:
         logger.critical("File is not a ZIP")
         sys.exit(-2)
@@ -287,13 +304,17 @@ def startup():
     if not os.path.exists(readme):
         logger.critical("Unable to determine base path correctly. Please make sure to run NZBHydra in the folder where its binary is located. Current base path: " + basePath)
         sys.exit(-1)
-    core = False
-    if os.path.isfile("core"):
-        args.java = "core"
-        core = True
-    if os.path.isfile("core.exe"):
-        args.java = "core.exe"
-        core = True
+
+    releaseType = determineReleaseType()
+    isWindows = platform.system().lower() == "windows"
+
+    if releaseType == ReleaseType.GENERIC:
+        args.java = "java"
+    else:
+        if isWindows:
+            args.java = "core.exe"
+        else:
+            args.java = "core"
 
     debugSwitchFile = os.path.join(args.datafolder, "DEBUG")
     if os.path.exists(debugSwitchFile):
@@ -303,9 +324,8 @@ def startup():
         console_logger.setLevel("DEBUG")
         logger.info("Setting wrapper log level to DEBUG")
 
-    isWindows = platform.system().lower() == "windows"
     libFolder = os.path.join(basePath, "lib")
-    if not core:
+    if releaseType == ReleaseType.GENERIC:
         if not os.path.exists(libFolder):
             logger.critical("Error: Lib folder %s not found. An update might've failed or the installation folder is corrupt", libFolder)
             sys.exit(-1)
@@ -374,38 +394,23 @@ def startup():
         logger.info("Removing superfluous M from XMX value " + xmx)
         xmx = xmx[:-1]
 
-    javaVersion = args.javaversion
-    if javaVersion is None and not core:
+    if releaseType == ReleaseType.GENERIC:
         javaVersion = getJavaVersion(args.java)
-    if core:
-        javaVersion = 17
+        if javaVersion < 17:
+            logger.critical("Error: Java 17 (not older, not newer) is required")
+            sys.exit(-1)
+
     gcLogFilename = (os.path.join(args.datafolder, "logs") + "/gclog-" + datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S") + ".log").replace("\\", "/")
     gcLogFilename = os.path.relpath(gcLogFilename, basePath)
 
-    gcArguments = []
-    if javaVersion < 9:
-        gcArguments = ["-Xloggc:" + gcLogFilename,
-                       "-XX:+PrintGCDetails",
-                       "-XX:+PrintGCTimeStamps",
-                       "-XX:+PrintTenuringDistribution",
-                       "-XX:+PrintGCCause",
-                       "-XX:+UseGCLogFileRotation",
-                       "-XX:NumberOfGCLogFiles=10",
-                       "-XX:GCLogFileSize=5M",
-                       ]
-    else:
-        gcArguments = [
-            "-Xlog:gc*:file=" + gcLogFilename + "::filecount=10,filesize=5000"]
+    gcArguments = [
+        "-Xlog:gc*:file=" + gcLogFilename + "::filecount=10,filesize=5000"]
 
-    java_arguments = ["-Xmx" + xmx + "M",
-                      "-DfromWrapper=true"
-                      ]
-    if not core:
+    java_arguments = ["-Xmx" + xmx + "M", "-DfromWrapper=true"]
+    if releaseType == ReleaseType.GENERIC:
         java_arguments.append("-XX:TieredStopAtLevel=1")
         java_arguments.append("-XX:+HeapDumpOnOutOfMemoryError")
         java_arguments.append("-XX:HeapDumpPath=" + os.path.join(args.datafolder, "logs"))
-    if javaVersion < 13:
-        java_arguments.append("-noverify")
     if logGc:
         java_arguments.extend(gcArguments)
     if args.debugport:
@@ -415,10 +420,10 @@ def startup():
     if args.debug:
         java_arguments.append("-Ddebug=true")
 
-    if not core:
-        arguments = [args.java] + java_arguments + ["-jar", escape_parameter(isWindows, jarFile)] + arguments
-    else:
+    if releaseType == ReleaseType.NATIVE:
         arguments = [args.java] + java_arguments + arguments
+    else:
+        arguments = [args.java] + java_arguments + ["-jar", escape_parameter(isWindows, jarFile)] + arguments
     commandLine = " ".join(arguments)
     logger.info("Starting NZBHydra main process with command line: %s in folder %s", commandLine, basePath)
     if hasattr(subprocess, 'STARTUPINFO'):
@@ -455,6 +460,21 @@ def startup():
         return process
     except Exception as e:
         logger.error("Unable to start process; make sure Java is installed and callable. Error message: " + str(e))
+
+
+def determineReleaseType():
+    if os.path.exists("ReleaseInfo.json"):
+        with open("ReleaseInfo.json", 'r') as f:
+            versionInfo = json.load(f)
+        releaseType = versionInfo["ReleaseType"]
+    elif os.path.exists("lib"):
+        releaseType = ReleaseType.GENERIC
+    else:
+        logger.critical(
+            "Unable to determine the release type. No VersionInfo.json file and no lib folder found")
+        sys.exit(-1)
+    logger.info("Determined release type: " + releaseType)
+    return releaseType
 
 
 def escape_parameter(is_windows, parameter):
@@ -538,7 +558,7 @@ if __name__ == '__main__':
     GracefulKiller()
     parser = argparse.ArgumentParser(description='NZBHydra 2')
     parser.add_argument('--java', action='store', help='Full path to java executable', default="java")
-    parser.add_argument('--javaversion', action='store', help='Force version of java  for which parameters java will be created', default=None)
+    parser.add_argument('--javaversion', action='store', help='Force version of java for which parameters java will be created', default=None)
     parser.add_argument('--debugport', action='store', help='Set debug port to enable remote debugging', default=None)
     parser.add_argument('--daemon', '-D', action='store_true', help='Run as daemon. *nix only', default=False)
     parser.add_argument('--pidfile', action='store', help='Path to PID file. Only relevant with daemon argument', default="nzbhydra2.pid")
@@ -561,6 +581,7 @@ if __name__ == '__main__':
 
     # Internal logic
     parser.add_argument('--restarted', action='store_true', default=False, help=argparse.SUPPRESS)
+    parser.add_argument('--update', action='store_true', default=False, help=argparse.SUPPRESS)
 
     args, unknownArgs = parser.parse_known_args()
     setupLogger()
@@ -582,6 +603,10 @@ if __name__ == '__main__':
     if os.path.exists(controlIdFilePath):
         os.remove(controlIdFilePath)
     doStart = True
+    if args.update:
+        logger.info("Executing update")
+        update()
+        sys.exit(0)
 
     if "--version" in unknownArgs or "--help" in unknownArgs:
         # no fancy shit, just start the file
